refactor: make add, remove follower procedure look  nice
