+++ src/meta-srv/src/metasrv/builder.rs
@@ -55,9 +55,8 @@ use crate::lease::MetaPeerLookupService;
 use crate::metasrv::{
     ElectionRef, Metasrv, MetasrvInfo, MetasrvOptions, SelectorContext, SelectorRef, TABLE_ID_SEQ,
 };
-use crate::procedure::add_region_follower::manager::{
-    AddRegionFollowerManager, Context as ArfContext,
-};
+use crate::procedure::region_follower::manager::RegionFollowerManager;
+use crate::procedure::region_follower::Context as ArfContext;
 use crate::procedure::region_migration::manager::RegionMigrationManager;
 use crate::procedure::region_migration::DefaultContextFactory;
 use crate::region::supervisor::{
@@ -347,7 +346,7 @@ impl MetasrvBuilder {
         );
 
         // add region follower manager
-        let add_region_follower_manager = Arc::new(AddRegionFollowerManager::new(
+        let add_region_follower_manager = Arc::new(RegionFollowerManager::new(
             procedure_manager.clone(),
             ArfContext {
                 table_metadata_manager: table_metadata_manager.clone(),
