+++ src/meta-srv/src/procedure.rs
@@ -18,7 +18,7 @@ use common_meta::leadership_notifier::LeadershipChangeListener;
 use common_procedure::ProcedureManagerRef;
 use snafu::ResultExt;
 
-pub mod add_region_follower;
+pub mod region_follower;
 pub mod region_migration;
 #[cfg(test)]
 mod tests;
